Second Commit after review

+++ project-management-api/project-management-api/src/main/java/com/HaikalArif/project_management_api/Service/UserService.java
@@ -1,9 +1,15 @@
 package com.HaikalArif.project_management_api.Service;
 
+import com.HaikalArif.project_management_api.Dto.UserDto;
+import com.HaikalArif.project_management_api.Exception.NotFoundException;
+import com.HaikalArif.project_management_api.Exception.UserAlreadyExistException;
+import com.HaikalArif.project_management_api.Mapper.UserMapper;
 import com.HaikalArif.project_management_api.Model.User;
 import com.HaikalArif.project_management_api.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,28 +20,50 @@ public class UserService {
 
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    public User createUser(UserDto userDto) {
+        if (existingByEmail(userDto.getEmail())) {
+            throw new UserAlreadyExistException("Email already used!");
+        }
+        userDto.setPassword(passwordEncoder.encode(userDto.getPassword())); // Encode password
+        User tempUser = UserMapper.toUserModel(userDto); // Convert User DTO to User model
+        User createUser = save(tempUser); // Save user details
+        if (createUser == null){
+            throw new NotFoundException("Required field is null!");
+        }
+        return createUser;
+    }
 
     // Create new user
-    public User saveOrUpdate(User user) {
+    public User save(User user) {
         return userRepository.save(user);
     }
-    // Try test for email check
-//    public User existsByEmail(User user) {
-//        return Optional<User> findExistingEmail = userRepository.findByEmail(user.getEmail());
-//    }
+
+    // Check existing email
     public boolean existingByEmail(String email) {
         return userRepository.existsByEmail(email);
     }
 
     // Get User by ID
-    public User getUserById(long id) {
+    public UserDto getUserById(long id) {
         Optional<User> getUser = userRepository.findById(id);
-        if (getUser.isPresent()) {
-            return getUser.get();
+        if (getUser.isEmpty()) {
+            throw new NotFoundException("Not Found User with ID " + id);
         }
-        else {
-            return null;
+        UserDto convertUser = UserMapper.toUserDto(getUser.get());
+        return convertUser;
+    }
+
+    public User updateUser(long id, UserDto userDto) {
+        userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
+        User tempUser = UserMapper.toUserModel(userDto);
+        User updatedUser = update(id, tempUser);
+        if (updatedUser == null) {
+            throw new NotFoundException("Unsuccessful Update! No User with Id "+ id);
         }
+        return updatedUser;
     }
 
     // Update User by ID
@@ -55,14 +83,22 @@ public class UserService {
     }
 
     // Delete user by ID
-    public void delete(long id) {
+    public String deleteById(long id) {
+        Optional<User> getUser = userRepository.findById(id);
+        if (getUser.isEmpty()) {
+            throw new NotFoundException("User not found with id " + id);
+        }
         userRepository.deleteById(id);
+        return "Successfully delete User ID " + id;
     }
 
     // Get all user
-    public List<User> getAllUser() {
+    public List<User> getAllUsers() {
         List<User> user = new ArrayList<User>();
         userRepository.findAll().forEach(user1 -> user.add(user1));
+        if (user.isEmpty()) {
+            throw new NotFoundException("User not found!");
+        }
         return user;
     }
 }
